requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
deta==1.0.1
Django==4.0
django-bootstrap-icons==0.7.5
django-bootstrap5==21.2
django-crispy-forms==1.13.0
frozenlist==1.2.0
h11==0.12.0
httpcore==0.14.3
idna==3.3
lxml==4.7.1
multidict==5.2.0
psycopg2-binary==2.9.2
pycparser==2.21
python-dotenv==0.19.2
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
yarl==1.7.2
